daemon: Fix arithmetic overflow
Fix arithmetic overflow.

+++ daemon/lock.c
@@ -305,7 +305,7 @@ retry_lock:
         DPRINTF(1,
             ("handle_lock(state->path.path='%s'): retry in %ldms\n",
             state->path.path, (long)poll_delay));
-        (void)delayxid(upcall->xid, 30+(poll_delay/1000));
+        (void)delayxid(upcall->xid, 30LL+((long long)poll_delay/1000LL));
         Sleep(poll_delay);
 
         goto retry_lock;
